Split ABI parsing into architecture-specific functionality
