+++ source/detail/parser/aarch64.hpp
@@ -1,0 +1,9 @@
+// Copyright 2013-2021 Lawrence Livermore National Security, LLC and other
+// Spack Project Developers. See the top-level COPYRIGHT file for details.
+//
+// SPDX-License-Identifier: (Apache-2.0 OR MIT)
+
+#pragma once
+
+namespace smeagle::aarch64 {
+}  // namespace smeagle::aarch64
