+++ source/detail/parser/ppc64le.cpp
@@ -1,0 +1,10 @@
+// Copyright 2013-2021 Lawrence Livermore National Security, LLC and other
+// Spack Project Developers. See the top-level COPYRIGHT file for details.
+//
+// SPDX-License-Identifier: (Apache-2.0 OR MIT)
+
+#include "ppc64le.hpp"
+
+namespace smeagle::ppc64le {
+
+}  // namespace smeagle::ppc64le
